+++ src/main/java/de/ghse/forum/api/request/CommentRequest.java
@@ -1,0 +1,11 @@
+package de.ghse.forum.api.request;
+
+import lombok.Data;
+
+@Data
+public class CommentRequest {
+    private String title;
+    private String text;
+    private String user_id;
+    private String post_id;
+}
