+++ src/main/java/de/ghse/forum/api/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package de.ghse.forum.api;
+package de.ghse.forum.api.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationResponse {
-  private String token;
+public class RegisterRequest {
+  private String user_name;
+  private String password;
 }
