+++ src/main/java/de/ghse/forum/api/response/AuthenticationResponse.java
@@ -1,4 +1,4 @@
-package de.ghse.forum.api;
+package de.ghse.forum.api.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationRequest {
-  private String user_name;
-  private String password;
+public class AuthenticationResponse {
+  private String token;
 }
