+++ src/main/java/de/ghse/forum/service/UserService.java
@@ -5,6 +5,8 @@ import de.ghse.forum.repository.UserRepository;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+
+import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -22,11 +24,12 @@ import org.springframework.stereotype.Service;
  * @since 1.0
  */
 @Service
+@RequiredArgsConstructor
 public class UserService {
   @Value("${page.size}")
   private int PAGE_SIZE;
 
-  @Autowired private UserRepository UserRepository;
+  private final UserRepository UserRepository;
 
   public void addUser(User user) {
     UserRepository.save(user);
